+++ realmFlipperPlugin/flipper-plugin-realm/src/index.tsx
@@ -1,16 +1,15 @@
 import {
   ConsoleSqlOutlined,
-  DatabaseOutlined, HistoryOutlined,
   SettingOutlined, TableOutlined
 } from '@ant-design/icons';
-import { StarOutlined } from '@ant-design/icons';
 
-import { Button, Radio, Typography, RadioChangeEvent, Input, Alert, Select, AutoComplete } from 'antd';
-const { Option } = AutoComplete;
+import { Button, Radio, Typography, RadioChangeEvent, AutoComplete } from 'antd';
 import { createState, Layout, PluginClient, Toolbar, usePlugin, useValue } from 'flipper-plugin';
-import React, { useState } from 'react';
+import React from 'react';
 import {useCallback} from 'react';
-type RealmPluginState = {
+import RealmQueryLanguage from "./pages/RealmQueryLanguage"
+
+export type RealmPluginState = {
   database: Number, 
   objects: Array<Object>,
   schemas: Array<Object>,
@@ -157,33 +156,7 @@ export function Component() {
   const onRQLClicked = useCallback(() => {
     instance.updateViewMode({viewMode: 'RQL'});
   }, [instance]);
-
-
-  const onTextChange = (event: String) => {
-    console.log("onTextChange", event);
-    instance.state.update(st => {
-      st.query = event
-    })
-  }
-
-  const onStar = () => {
-    console.log("onStar");
-    const state = instance.state.get()
-    if (!state.queryFavourites.includes(state.query)) {
-      instance.state.update(st => {
-        st.queryFavourites = [...st.queryFavourites, st.query]
-      })
-    }
-  }
-  const render2 = (query: String, id: number) => (
-    {
-      label: query,
-      value: query,
-      key: id
-    }
-    // <Option value={query} key={id}>query</Option>
-  );
-  console.log(state.viewMode)
+  // console.log(state.viewMode)
 
   return (
     <Layout.Container grow>
@@ -222,44 +195,7 @@ export function Component() {
             <div>schemas tab</div>
       : null} 
       {state.viewMode === 'RQL' ? (<>
-        <Input.Group compact>
-          <AutoComplete style={{ width: 'calc(100% - 115px)' }} 
-            placeholder="Enter a query to filter the data"
-            onSearch={onTextChange} id="msgbox"
-            onChange={onTextChange}
-            onKeyUp={(ev) => {
-              if (ev.key == 'Enter')
-                instance.executeQuery()
-            }}
-            // onPressEnter={instance.executeQuery}
-            allowClear
-            showSearch
-            options={[{
-              label: 'History',
-              options: state.queryHistory.map((val, id) => render2(val, 2 * id)).reverse()
-            },
-            {
-              label: 'Favourites',
-              options: state.queryFavourites.map((val, id) => render2(val, 2 * id + 1)).reverse()
-            }]}
-            backfill={true}
-            >
-          </AutoComplete>
-          <Button type="primary" onClick={instance.executeQuery} title="executeButton">Execute</Button>
-          <Button icon={<StarOutlined />} onClick={onStar}></Button>
-        </Input.Group>
-        {state.objects.map((obj) => {
-          // @ts-ignore
-          return (<div key={ obj._id}>{JSON.stringify(obj)}</div>)
-        })}
-      {state.errorMsg ? (
-        <Alert
-          message="Error"
-          description={state.errorMsg}
-          type="error"
-          showIcon
-          banner={true}
-      />): null}
+        <RealmQueryLanguage instance={instance}></RealmQueryLanguage>
       </>
       ) : null}
     </Layout.Container>
